// shared/types/infinity-reputation-engine.ts
export enum ReputationType {
  CONTRIBUTION = 'contribution',
  TRUST = 'trust',
  GOVERNANCE = 'governance',
  SOCIAL = 'social',
  ECONOMIC = 'economic'
}

export enum MeasurementMethod {
  AUTOMATIC = 'automatic',
  PEER_VALIDATION = 'peer_validation',
  AI_ANALYSIS = 'ai_analysis',
  COMMUNITY_VOTE = 'community_vote',
  HYBRID = 'hybrid'
}

export enum DecayFunction {
  LINEAR = 'linear',
  EXPONENTIAL = 'exponential',
  LOGARITHMIC = 'logarithmic',
  STEP_FUNCTION = 'step_function',
  SIGMOID = 'sigmoid',
  CUSTOM = 'custom'
}
// ...
export enum InterventionType {
  AUTOMATIC_ADJUSTMENT = 'automatic_adjustment',
  PARAMETER_TUNING = 'parameter_tuning',
  EMERGENCY_INTERVENTION = 'emergency_intervention',
  GOVERNANCE_OVERRIDE = 'governance_override',
  ECONOMIC_STIMULUS = 'economic_stimulus',
  SOCIAL_INTERVENTION = 'social_intervention',
  REPUTATION_RESET = 'reputation_reset'
}
// ...
export enum TransactionType {
  EARNED = 'earned',
  STAKED = 'staked',
  BURNED = 'burned',
  TRANSFERRED = 'transferred',
  DECAYED = 'decayed',
  CONFISCATED = 'confiscated',
  REWARDED = 'rewarded',
  PENALIZED = 'penalized'
}